NEW: First Commit
This first commit set up a RPG like project
using Unity3D version 5.3.1. This project is
based on the SurvivalShooter tutorial by Unity.

Unity release on Ubuntu is still causing issues but
the project shouldn't be affected.

The scripts located in the Assets/Scripts folder
have been adapted from the tutorial in order to
get a RPG like battle scene. Indent and consistency
have to be fixed in future commits.